requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic-settings==2.4.0
python-multipart==0.0.9
